pyproject.toml
[tool.poetry]
name = "best-parser-bot"
version = "1.0.0"
description = "Rule34.xxx parser."
authors = ["Fedor <fed.degtjarev@gmail.com>"]
readme = "README.md"
packages = [{include = "bot"}]

[tool.poetry.dependencies]
python = "^3.11"
python_telegram_bot = "==20.2"
requests = "^2.28.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
